requirements.txt
pandas==2.2.3
playwright==1.52.0
openpyxl>=3.1.0
Flask>=3.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
flask-mail>=0.9.1
redis>=5.0.0
rq>=1.15.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
